activation features
This commit adds single clause activation features to Gini.

There is a new inter.S interface, Activatable which asks the
caller to use 'Activate()' instead of 'Add(0)' to generate an
activatable clause.  A literal is returned, which if assumed,
activates the clause.

There is a Deactivate method which removes the activation
clause and the associated learned clauses.  Deactivate then
frees the associated activation literal for future Activations.

When using Activations, the caller must not inadvertantly use
activation literals.  So once activations are in use, all new
literals should come from calls to `Lit()` and not by explicit
variable indexing.