aiger bugfix and usability improvements
This commit fixes a bug in the aiger ascii output for latch values and
adds some usability/performance improvements

For usability, we have added a Write method to gini to have it dump
a dimacs CNF to a writer.

For performance, we have made it so that more learned clauses could be
derived in incremental mode with test scopes when the last solve gave
unsat.